requirements.txt
beautifulsoup4==4.12.2
keyboard==0.13.5
mss==9.0.1
pandas==2.0.2
pyautogui==0.9.54
pynput==1.7.6
PyYAML==6.0
selenium==4.9.1
